Added safe method to get cache path string, separate function to create environment file, type annotations for parameters and returns, , removed unnecessary checks.

+++ src/NIHData/cache/env.py
@@ -2,8 +2,9 @@ from pathlib import Path
 from NIHData.errors import CacheDoesNotExist, EnvFileDoesNotExist, EnvVarDoesNotExist
 
 NIH_CACHE_ENV_VAR_NAME = 'NIH_DATA_CACHE_DIR'
-NIH_CACHE_ENV_FILE_NAME = f".nih_data_cache.env"
+NIH_CACHE_ENV_FILE_NAME = ".nih_data_cache.env"
 NIH_CACHE_ENV_FILE_PATH = Path.home() / NIH_CACHE_ENV_FILE_NAME
+NIH_CACHE_DEFAULT_PATH = Path.home() / ".nih_data_cache"
 
 
 def _read_env_file():
@@ -19,17 +20,38 @@ def _read_env_file():
     return mapping
 
 
-def get_cache_path():
+def _get_cache_path_string() -> str | None:
     """
-    Reads environment variable and returns the path to the cache directory, if it exists.
+    Safely returns the path to the cache (as a string) without raising.
+    """
+    try:
+        env_data = _read_env_file()
+        return env_data.get(NIH_CACHE_ENV_VAR_NAME, None)
+    except EnvFileDoesNotExist as e:
+        print(e)
+        return None
+
+
+def setup_env_file(cache_path: Path) -> None:
+    print(f"Setting {NIH_CACHE_ENV_VAR_NAME}={cache_path}.")
+    lines = (
+        f"# This file holds the environment variable for the NIH Data Cache.",
+        f"{NIH_CACHE_ENV_VAR_NAME}={cache_path}"  # sets env var
+    )
+    with open(NIH_CACHE_ENV_FILE_PATH, mode='w', encoding='utf-8') as envfile:
+        envfile.write("\n".join(lines))
+
+
+def get_cache_path() -> Path:
+    """
+    Reads environment variable and returns the cache_path to the cache directory, if it exists.
 
     :return: Path to the cache directory.
     :raises EnvVarDoesNotExist: Raised if the environment variable is not found in the .env mapping.
     :raises EnvFileDoesNotExist: Raised if the environment variable file has not been created yet.
     """
     try:
-        env_data = _read_env_file()
-        if cache_path_str := env_data.get(NIH_CACHE_ENV_VAR_NAME, None):
+        if cache_path_str := _get_cache_path_string():
             cache_path = Path(cache_path_str)
             if not cache_path.exists():
                 raise CacheDoesNotExist("Cache directory does not exist.")
@@ -42,14 +64,13 @@ def get_cache_path():
         raise e
 
 
-def create_cache_directory(path: str | Path | None = None):
+def create_cache_directory(cache_path: str | Path | None = None) -> Path:
     """
-    Defaults to 'Path.home() / .nih_data_cache' if no direct path is provided to generate the cache.
+    Defaults to 'Path.home() / .nih_data_cache' if no direct cache_path is provided to generate the cache.
 
     Also checks to see if an .nih_env_cache.env file exists in the Path.home() dir; if not, creates one.
     """
-
-    nih_data_cache_path = Path(path) if path else Path.home() / '.nih_data_cache'
+    nih_data_cache_path = Path(cache_path) if cache_path else NIH_CACHE_DEFAULT_PATH
 
     if not nih_data_cache_path.exists():
         print(f"No cache directory found at {nih_data_cache_path}. Creating...")
@@ -70,7 +91,7 @@ def create_cache_directory(path: str | Path | None = None):
     return nih_data_cache_path
 
 
-def delete_cache_directory():
+def delete_cache_directory() -> None:
     print("Deleting files in cache directory, and environment variable.")
     mapping = _read_env_file()
     if NIH_CACHE_ENV_VAR_NAME not in mapping:
@@ -82,38 +103,15 @@ def delete_cache_directory():
                          "This is a universal short circuit; please check your environment setup before "
                          "proceeding.")
 
-    for file in nih_data_cache_path.iterdir():
-        print(f"[FILE] Deleting {file}")
+    for n, file in enumerate(nih_data_cache_path.iterdir(), start=1):
+        print(f"[FILE-{n}] Deleting {file} from cache")
         file.unlink()
 
     print(f"[DIRECTORY] Deleting {nih_data_cache_path}")
     nih_data_cache_path.rmdir()
-    print(f"[ENV_FILE] Deleting {NIH_CACHE_ENV_FILE_PATH}")
-    NIH_CACHE_ENV_FILE_PATH.unlink()
+    print("[SUCCESS] Cache successfully deleted.")
 
-    print("[VALIDATION] Running checks...")
-
-    checks = enumerate((
-        (not nih_data_cache_path.exists(),
-         "NIH data cache successfully deleted.",
-         "NIH data cache not deleted. Check environment variable."
-         ),
-        (not NIH_CACHE_ENV_FILE_PATH.exists(),
-         "NIH environment variable successfully deleted.",
-         f"NIH environment variable not deleted. Please review what is located at {NIH_CACHE_ENV_FILE_PATH}."
-         )
-    ), start=1)
-
-    errors = []
-
-    for n, check in checks:
-        valid, success_message, error = check
-        if valid:
-            print(f"[{n}] {success_message}")
-        else:
-            errors.append(error)
-
-    if errors:
-        raise ValueError(errors)
-
-    print("[SUCCESS] Cache successfully deleted!")
+    if NIH_CACHE_ENV_FILE_PATH.exists():
+        print(f"[ENV_FILE] Deleting {NIH_CACHE_ENV_FILE_PATH}")
+        NIH_CACHE_ENV_FILE_PATH.unlink()
+        print("[SUCCESS] Environment file successfully deleted.")
